E-CommerceWeb/ecommerce_backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
bcrypt==3.2.0
mysqlclient==2.1.1
django-elasticsearch-dsl==9.0
elasticsearch==9.2.0
